requirements.txt
aiogram==3.11.0
openpyxl==3.1.2
python-dotenv==1.0.1